+++ src/test/kotlin/me/bossm0n5t3r/leetcode/jumpGame2/JumpGame2Test.kt
@@ -1,10 +1,10 @@
-package jumpGame2
+package me.bossm0n5t3r.leetcode.jumpGame2
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 internal class JumpGame2Test {
-    private val sut = JumpGame2()
+    private val sut = JumpGame2.Solution()
 
     private data class Example(
         val nums: IntArray,
